elastic_panel/test/requirement.txt
django-debug-toolbar
elasticsearch
nose